Explicit cast from num_vtxbufs to int to avoid warning (should be fine, likely not going to have >2147483647 fields anytime soon)

+++ src/core/kernels/packing.cuh
@@ -109,7 +109,8 @@ kernel_partial_pack_data(const VertexBufferArray vba, const int3 vba_start, cons
     const size_t vtxbuf_offset = dims.x * dims.y * dims.z;
 
     //#pragma unroll
-    for (int i = 0; i < num_vtxbufs; ++i) {
+    // Note explicit cast size_t to int
+    for (int i = 0; i < (int)num_vtxbufs; ++i) {
         int vtxbuf_id                          = vtxbufs[i];
         packed[packed_idx + i * vtxbuf_offset] = vba.in[vtxbuf_id][unpacked_idx];
     }
@@ -143,7 +144,8 @@ kernel_partial_unpack_data(const AcRealPacked* packed, const int3 vba_start, con
     const size_t vtxbuf_offset = dims.x * dims.y * dims.z;
 
     //#pragma unroll
-    for (int i = 0; i < num_vtxbufs; ++i) {
+    // Note explicit cast size_t to int
+    for (int i = 0; i < (int)num_vtxbufs; ++i) {
         int vtxbuf_id                   = vtxbufs[i];
         vba.in[vtxbuf_id][unpacked_idx] = packed[packed_idx + i * vtxbuf_offset];
     }
